pyproject.toml
[project]
name = "coda"
version = "0.1.0"
description = "AI-powered coding agent with Tokyo Night theme"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.60.0",
    "rich>=13.7.0",
    "openai>=1.0.0",
    "gitpython>=3.1.0",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
coda = "coda.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coda"]

[tool.uv]
# uv will manage a local environment automatically on uv run
package = true